Removed acceptance testing code.
The only way to run acceptance tests on different machines at this
moment seems to be running QEMU inside another virtual machine as it
does not rely on virtualization hardware features to emulate the guest.

Doing this however would make the acceptance tests almost unusable as
no one would run QEMU in their environment.

Decided to remove CI acceptance test harness for now and focus on
improving installation of software on guests.